# docs/api_spec.yaml
openapi: 3.0.0
info:
  title: 儿童教育AI系统API
  description: 为儿童教育AI系统提供的RESTful API接口
  version: 1.0.0
  contact:
    name: 开发团队
    email: dev@edu-ai-system.com

servers:
  - url: http://localhost:5000/api
    description: 本地开发环境
  - url: https://api.edu-ai-system.com
    description: 生产环境

paths:
  /health:
    get:
      summary: 健康检查
      description: 检查API服务是否正常运行
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthCheckResponse'
        '500':
          description: 服务异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /chat:
    post:
      summary: 聊天接口
      description: 与AI助手进行对话
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: 成功响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /user/{user_id}/history:
    get:
      summary: 获取用户对话历史
      description: 获取指定用户的对话历史记录
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: 用户ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
          description: 返回记录数量限制
      responses:
        '200':
          description: 成功获取历史记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserHistoryResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /stats:
    get:
      summary: 获取系统统计信息
      description: 获取系统的统计信息，包括用户、对话、安全等数据
      responses:
        '200':
          description: 成功获取统计信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemStatsResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /safety/violations:
    get:
      summary: 获取安全违规记录
      description: 获取最近的安全违规记录
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: 返回记录数量限制
      responses:
        '200':
          description: 成功获取安全违规记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafetyViolationsResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /search:
    get:
      summary: 搜索对话记录
      description: 根据关键词搜索对话记录
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
          description: 搜索关键词
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: 返回记录数量限制
      responses:
        '200':
          description: 成功搜索到记录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: 用户输入的文本
          example: "你好，我想学习数学"
        user_id:
          type: string
          description: 用户ID
          default: "default_user"
          example: "child_001"
        emotion:
          type: string
          description: 用户情感状态
          example: "happy"
      additionalProperties: false

    ChatResponse:
      type: object
      required:
        - status
        - response
        - agent
      properties:
        status:
          type: string
          description: 响应状态
          example: "success"
        response:
          type: string
          description: AI助手的回复
          example: "你好！我很乐意帮助你学习数学。你想了解哪个方面的数学知识呢？"
        agent:
          type: string
          description: 处理请求的Agent名称
          example: "EduAgent"

    HealthCheckResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          description: 服务状态
          example: "healthy"
        message:
          type: string
          description: 状态描述信息
          example: "Education AI System is running"

    ConversationHistoryItem:
      type: object
      required:
        - user_id
        - user_input
        - agent_response
        - agent_name
      properties:
        id:
          type: integer
          description: 记录ID
        user_id:
          type: string
          description: 用户ID
        user_input:
          type: string
          description: 用户输入
        agent_response:
          type: string
          description: AI助手回复
        agent_name:
          type: string
          description: 处理Agent名称
        safety_check:
          type: string
          description: 安全检查结果
        timestamp:
          type: string
          format: date-time
          description: 时间戳

    UserHistoryResponse:
      type: object
      required:
        - status
        - history
      properties:
        status:
          type: string
          description: 响应状态
          example: "success"
        history:
          type: array
          items:
            $ref: '#/components/schemas/ConversationHistoryItem'

    SafetyViolationItem:
      type: object
      required:
        - user_input
        - agent_response
        - violation_reason
        - agent_name
      properties:
        id:
          type: integer
          description: 记录ID
        user_id:
          type: string
          description: 用户ID
        user_input:
          type: string
          description: 用户输入
        agent_response:
          type: string
          description: AI助手回复
        violation_reason:
          type: string
          description: 违规原因
        agent_name:
          type: string
          description: 处理Agent名称
        timestamp:
          type: string
          format: date-time
          description: 时间戳

    SafetyViolationsResponse:
      type: object
      required:
        - status
        - violations
      properties:
        status:
          type: string
          description: 响应状态
          example: "success"
        violations:
          type: array
          items:
            $ref: '#/components/schemas/SafetyViolationItem'

    SystemStatsResponse:
      type: object
      required:
        - status
        - stats
      properties:
        status:
          type: string
          description: 响应状态
          example: "success"
        stats:
          type: object
          description: 系统统计信息
          properties:
            users:
              type: object
              properties:
                total_users:
                  type: integer
                active_users:
                  type: integer
                active_user_rate:
                  type: number
                  format: float
                grade_distribution:
                  type: object
                  additionalProperties:
                    type: integer
            conversations:
              type: object
              properties:
                total_conversations:
                  type: integer
                agent_distribution:
                  type: object
                  additionalProperties:
                    type: integer
                daily_conversations:
                  type: object
                  additionalProperties:
                    type: integer
            safety:
              type: object
              properties:
                total_violations:
                  type: integer
                violation_distribution:
                  type: object
                  additionalProperties:
                    type: integer
                agent_violations:
                  type: object
                  additionalProperties:
                    type: integer

    SearchResultItem:
      type: object
      required:
        - user_id
        - user_input
        - agent_response
        - agent_name
      properties:
        id:
          type: integer
          description: 记录ID
        user_id:
          type: string
          description: 用户ID
        user_input:
          type: string
          description: 用户输入
        agent_response:
          type: string
          description: AI助手回复
        agent_name:
          type: string
          description: 处理Agent名称
        safety_check:
          type: string
          description: 安全检查结果
        timestamp:
          type: string
          format: date-time
          description: 时间戳

    SearchResponse:
      type: object
      required:
        - status
        - results
        - count
      properties:
        status:
          type: string
          description: 响应状态
          example: "success"
        results:
          type: array
          items:
            $ref: '#/components/schemas/SearchResultItem'
        count:
          type: integer
          description: 搜索结果总数

    ErrorResponse:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: string
          description: 错误状态
          example: "error"
        message:
          type: string
          description: 错误信息
          example: "请求参数错误"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - bearerAuth: []

tags:
  - name: health
    description: 健康检查相关接口
  - name: chat
    description: 聊天相关接口
  - name: user
    description: 用户相关接口
  - name: stats
    description: 统计相关接口
  - name: safety
    description: 安全相关接口
  - name: search
    description: 搜索相关接口